+++ applications/fishing-map/src/features/workspace/context-areas/ContextAreaLayerPanel.tsx
@@ -66,7 +66,9 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
   if (!dataset) {
     return <DatasetNotFound dataview={dataview} />
   }
-  const title = isCustomUserLayer ? dataset?.name || dataset?.id : t(`datasets:${dataset?.id}.name`)
+  const title = isCustomUserLayer
+    ? dataset?.name || dataset?.id
+    : t(`datasets:${dataset?.id}.name` as any)
   const TitleComponent = (
     <h3 className={cx(styles.name, { [styles.active]: layerActive })} onClick={onToggleLayerActive}>
       {title}
@@ -77,7 +79,7 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
   const datasetError = dataset.status === DatasetStatus.Error
   let infoTooltip = isCustomUserLayer
     ? dataset?.description
-    : t(`datasets:${dataset?.id}.description`)
+    : t(`datasets:${dataset?.id}.description` as any)
   if (datasetImporting) {
     infoTooltip = t('dataset.importing', 'Dataset is being imported')
   }